chore(scap): move scap_ns_socket_list to scap_linux_int.h
