+++ userspace/libscap/linux/scap_linux_int.h
@@ -17,8 +17,22 @@ limitations under the License.
 
 #pragma once
 
-struct scap_ns_socket_list;
+#include <stdint.h>
+
+#include "uthash.h"
+
+typedef struct scap_fdinfo scap_fdinfo;
+
+struct scap_ns_socket_list
+{
+	int64_t net_ns;
+	scap_fdinfo* sockets;
+	UT_hash_handle hh;
+};
+
 
 // read all sockets and add them to the socket table hashed by their ino
 int32_t scap_fd_read_sockets(char* procdir, struct scap_ns_socket_list* sockets, char *error);
 void scap_fd_free_ns_sockets_list(struct scap_ns_socket_list** sockets);
+// read the file descriptors for a given process directory
+int32_t scap_fd_scan_fd_dir(scap_t* handle, char * procdir, scap_threadinfo* pi, struct scap_ns_socket_list** sockets_by_ns, uint64_t* num_fds_ret, char *error);
